password-store: add support
This shall be used no only as a local keyring for users as they see fit,
but also (and more importantly) for mutt and openvpn. With these scripts
we rely on a specific gnupg keyring (aside from the local .gnupg
keyring), and allow for a centralized, single source of thruth password-store,
using 'password-store's git interface.

Signed-off-by: Joao Eduardo Luis <joao@wipwd.org>